feat: hystrix integration, datatypes changing

+++ src/main/java/com/nesterrovv/vpntoken/repository/TokenRepository.java
@@ -3,6 +3,6 @@ package com.nesterrovv.vpntoken.repository;
 import com.nesterrovv.vpntoken.entity.Token;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
 
-public interface TokenRepository extends R2dbcRepository<Token, Long> {
+public interface TokenRepository extends R2dbcRepository<Token, Integer> {
 
 }
